HW3_Instance-Segmentation/requirements.txt
matplotlib==3.10.1
numpy==2.2.5
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
Pillow==11.2.1
pycocotools==2.0.8
scikit_learn==1.6.1
skimage==0.0
torchmetrics==1.7.0
